Test that twikified can talk with a (fake) twikifier node app
Simulates the other end of the sockect connection that the node app twikifier would listen on.

At the moment the connection just sends back some canned data.
It needs to be modified to continuously serve requests from the tests and behave differently for certain requests to test the socket robustness (e.g. close connection).